+++ ruoyi-paychology/src/main/java/com/ruoyi/course/mapper/CourCourseBannerConfigMapper.java
@@ -18,7 +18,7 @@ public interface CourCourseBannerConfigMapper
      * @param id 课程banner配置主键
      * @return 课程banner配置
      */
-    public CourCourseBannerConfig selectCourCourseBannerConfigById(Long id);
+    public CourCourseBannerConfig selectCourCourseBannerConfigById(Integer id);
 
     /**
      * 查询课程banner配置列表
@@ -50,7 +50,7 @@ public interface CourCourseBannerConfigMapper
      * @param id 课程banner配置主键
      * @return 结果
      */
-    public int deleteCourCourseBannerConfigById(Long id);
+    public int deleteCourCourseBannerConfigById(Integer id);
 
     /**
      * 批量删除课程banner配置
@@ -58,5 +58,5 @@ public interface CourCourseBannerConfigMapper
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deleteCourCourseBannerConfigByIds(Long[] ids);
+    public int deleteCourCourseBannerConfigByIds(Integer[] ids);
 }
